修改了后端授权的流程，增加了安全性

+++ ai-chat-web/ai-chat/src/main/java/com/k8/vo/TokenRefreshVO.java
@@ -16,5 +16,5 @@ public class TokenRefreshVO {
     private String accessToken;
     private String refreshToken;
     private String tokenType;
-    private Integer expiresIn;
+    private Long expiresIn;
 }
